# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VCanCare API",
        "description": "VCanCare Backend APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/nurse/onboarding/step-data/{step}": {
            "get": {
                "tags": [
                    "Nurse Onboarding"
                ],
                "summary": "Get Step Data",
                "operationId": "getStepData",
                "parameters": [
                    {
                        "name": "step",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}